+++ app/Http/Controllers/ProcessorsController.php
@@ -2,35 +2,58 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Processors;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Validator;
 
 class ProcessorsController extends Controller
 {
+    // Get request
     public function index()
     {
        return Processors::all();
     }
 
+    // Post for creating
     public function store(Request $request)
     {
-        $fields = $request->validate([
+        $fields = Validator::make($request->all(), [
             'processor_name'    => 'required|string',
             'brand'             => 'required|string',
             'socket_type'       => 'required|string',
             'tdp_wattage'       => 'required|integer',
             'base_clock_speed'  => 'required|numeric',
             'max_clock_speed'   => 'required|numeric',
-
         ]);
 
-        $processors = Processors::create($fields);
+        if($fields->fails()){
+            return response()->json([
+                'message'=>'All fields are mandatory',
+                'error'=>$fields->errors()
+            ],422);
+        }
 
-        return response()->json(['message'=> 'Created successful'],200);
+        $validatedData = $fields->validated();
+        $processors = Processors::create($validatedData);
+
+        return response()->json([
+            'status'=> true,
+            'message'=> 'Created successful',
+            'data' => $processors
+        ],200);
     }
 
-    public function show($id)
+    // Get request by specific ID
+    public function show($processor_id)
     {
-        $processors = Processors::findOrFail($id);
+        $processors = Processors::find($processor_id);
+        if (!$processors) {
+            return response()->json([
+                'status' => false,
+                'message' => 'Processor data not found'
+            ], 404);
+        }
+
         return response()->json([
             'status' => true,
             'message' => 'Processors data found successfully',
@@ -38,35 +61,56 @@ class ProcessorsController extends Controller
         ], 200);
     }
 
-
-
+    // Update
 public function update(Request $request, $processors)
 {
-    $fields = $request->validate([
-            'processor_name'    => 'required|string',
-            'brand'             => 'required|string',
-            'socket_type'       => 'required|string',
-            'tdp_wattage'       => 'required|integer',
-            'base_clock_speed'  => 'required|numeric',
-            'max_clock_speed'   => 'required|numeric',
+    $fields = Validator::make($request->all(),[
+        'processor_name'    => 'required|string',
+        'brand'             => 'required|string',
+        'socket_type'       => 'required|string',
+        'tdp_wattage'       => 'required|integer',
+        'base_clock_speed'  => 'required|numeric',
+        'max_clock_speed'   => 'required|numeric',
     ]);
 
+    if($fields->fails()){
+        return response()->json([
+            'message'=>'All fields are mandatory',
+            'error'=>$fields->errors()
+        ],422);
+    }
+
     $processors = Processors::find($processors);
+    if (!$processors) {
+        return response()->json([
+            'status' => false,
+            'message' => 'Processor data not found'
+        ], 404);
+    }
 
-    $processors->update($fields);
+    $processors->update($fields->validated());
 
-    return response()->json($processors, 201);
+    return response()->json([
+        'status' => true,
+        'message' => 'Processor Data Updated Successfully',
+        'data' => $processors], 201);
 }
 
-
-
+    // Delete requests by specific ID
     public function destroy($processors)
     {
         $processors = Processors::find($processors);
+        if (!$processors) {
+            return response()->json([
+                'status' => false,
+                'message' => 'Processors data not found'
+            ], 404);
+        }
         $processors->delete();
 
         return response()->json([
-            'message' => 'Processors data deleted successfully'
+            'message' => 'Processors data deleted successfully',
+            'data' => $processors
         ], 200);
     }
 }
