+++ app/Http/Controllers/RamsController.php
@@ -2,34 +2,57 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Rams;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Validator;
 
 class RamsController extends Controller
 {
+    // Get request
     public function index()
     {
        return Rams::all();
     }
 
+    // Post for creating
     public function store(Request $request)
     {
-        $fields = $request->validate([
+        $fields = Validator::make($request->all(),[
             'ram_name'         => 'required|string',
             'brand'            => 'required|string',
             'ram_type'         => 'required|string',
             'ram_capacity_gb'  => 'required|integer',
             'ram_speed_mhz'    => 'required|integer',
-
         ]);
 
-        $rams = Rams::create($fields);
+        if($fields->fails()){
+            return response()->json([
+                'message'=>'All fields are mandatory',
+                'error'=>$fields->errors()
+            ],422);
+        }
 
-        return response()->json(['message'=> 'Created successful'],200);
+        $validatedData = $fields->validated();
+        $rams = Rams::create($validatedData);
+
+        return response()->json([
+            'status'=> true,
+            'message'=> 'Created successful',
+            'data' => $rams
+        ],200);
     }
 
-    public function show($id)
+    // Get request by specific ID
+    public function show($ram_id)
     {
-        $rams = Rams::findOrFail($id);
+        $rams = Rams::find($ram_id);
+        if (!$rams) {
+            return response()->json([
+                'status' => false,
+                'message' => 'Ssd data not found'
+            ], 404);
+        }
+
         return response()->json([
             'status' => true,
             'message' => 'Ram data found successfully',
@@ -37,31 +60,57 @@ class RamsController extends Controller
         ], 200);
     }
 
-
+// Update
 public function update(Request $request, $rams)
 {
-    $fields = $request->validate([
-            'ram_name'         => 'required|string',
-            'brand'            => 'required|string',
-            'ram_type'         => 'required|string',
-            'ram_capacity_gb'  => 'required|integer',
-            'ram_speed_mhz'    => 'required|integer',
+    $fields = Validator::make($request->all(),[
+        'ram_name'         => 'required|string',
+        'brand'            => 'required|string',
+        'ram_type'         => 'required|string',
+        'ram_capacity_gb'  => 'required|integer',
+        'ram_speed_mhz'    => 'required|integer',
     ]);
+
+    if($fields->fails()){
+        return response()->json([
+            'message'=>'All fields are mandatory',
+            'error'=>$fields->errors()
+        ],422);
+    }
 
     $rams = Rams::find($rams);
 
-    $rams->update($fields);
+    if (!$rams) {
+        return response()->json([
+            'status' => false,
+            'message' => 'Ssd data not found'
+        ], 404);
+    }
 
-    return response()->json($rams, 201);
-}
+    $rams->update($fields->validated());
 
+    return response()->json([
+        'status' => true,
+        'message' => 'Ram Data Updated Successfully',
+        'data' => $rams], 201);
+    }
+
+    // Delete requests by specific ID
     public function destroy($rams)
     {
         $rams = Rams::find($rams);
+        if (!$rams) {
+            return response()->json([
+                'status' => false,
+                'message' => 'Ram data not found'
+            ], 404);
+        }
+
         $rams->delete();
 
         return response()->json([
-            'message' => 'Ram data deleted successfully'
+            'message' => 'Ram Data Deleted Successfully',
+            'data' => $rams
         ], 200);
     }
 }
